requirements.txt
alembic==1.0.0
APScheduler==3.5.3
certifi==2018.8.24
chardet==3.0.4
click==6.7
environs==4.1.0
Flask==1.0.2
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
get==1.0.3
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonify==0.5
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.16.3
post==1.0.2
psycopg2==2.7.3.1
public==1.0.3
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
query-string==1.0.2
redis==3.0.1
requests==2.20.0
rq==0.13.0
six==1.11.0
SQLAlchemy==1.2.12
SQLAlchemy-Utils==0.33.5
timedelta==0.0.1
todoist-python==7.0.18
tzlocal==1.5.1
urllib3==1.23
Werkzeug==0.14.1
